Remove dead PumpAsync code
